requirements.txt
fastapi_mail
fastapi
pymongo
uvicorn
pytz
pydantic
pandas
cryptography
openai
bcrypt
google-cloud-storage
python-pptx
apscheduler
mongoengine
loguru
streamlit
google-api-python-client
